requirements.txt

socket
threading
pyautogui
pickle
zlib
tkinter
Pillow
